LayoutTests/storage/storageinfo-missing-arguments.html

<!DOCTYPE html>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script>

description('Ensures that we get exceptions when we do not include required arguments for webkitStorageInfo methods.');

if (window.webkitStorageInfo) {
    var notEnoughArguments = '"TypeError: Not enough arguments"';
    shouldThrow('webkitStorageInfo.requestQuota()', notEnoughArguments);
    shouldThrow('webkitStorageInfo.requestQuota(webkitStorageInfo.TEMPORARY)', notEnoughArguments);
    shouldThrow('webkitStorageInfo.queryUsageAndQuota()', notEnoughArguments);
} else
    debug("This test requires window.webkitStorageInfo.");

</script>
<script src="../fast/js/resources/js-test-post.js"></script>
